+++ migrations/20161205190829-create-commission.js
@@ -13,13 +13,13 @@ module.exports = {
       //TODO: tiered commissions by default?
       commissionType: Sequelize.ENUM('none', 'traditional', 'tieredTotal', 'tieredMarginal'),
       tier0price: Sequelize.INTEGER,
-      tier0Commission: Sequelize.INTEGER,
+      tier0Commission: Sequelize.DECIMAL(10,2),
       tier1Price: Sequelize.INTEGER,
-      tier1Commission: Sequelize.INTEGER,
+      tier1Commission: Sequelize.DECIMAL(10,2),
       tier2Price: Sequelize.INTEGER,
-      tier2Commission: Sequelize.INTEGER,
+      tier2Commission: Sequelize.DECIMAL(10,2),
       tier3Price: Sequelize.INTEGER,
-      tier3Commission: Sequelize.INTEGER,
+      tier3Commission: Sequelize.DECIMAL(10,2),
       createdAt: {
         allowNull: false,
         type: Sequelize.DATE
